+++ src/interfaces/interface.ts
